Added checkpoint shadow, optimized shader function && shaders

+++ project/data/road.frag
@@ -32,42 +32,42 @@ varying highp vec3 tangentVarying;    // tangent in world space
 
 void main()
 {
-    lowp vec4 ambientResult = vec4(Ka * Ia, 1.0);
+    highp vec4 ambientResult = vec4(Ka * Ia, 1.0);
     
-    mediump vec4 pos = posVarying;
-    mediump vec3 n = normalize(normalVarying);
+    highp vec4 pos = posVarying;
+    highp vec3 n = normalize(normalVarying);
     
     // TODO: calculate tbn matrix, ensure it's orthonormal
-    mediump vec3 t = normalize(tangentVarying);
+    highp vec3 t = normalize(tangentVarying);
     t = normalize(t-n*dot(n,t));
-    mediump vec3 b = normalize(cross(t, n));
+    highp vec3 b = normalize(cross(t, n));
     
-    mediump mat3 tbn = mat3(t, b, n);
+    highp mat3 tbn = mat3(t, b, n);
     
     // TODO: read and correctly transform normals from normal map, then use them for lighting
-    mediump vec3 bump = texture2D(NormalMap, texCoordVarying.st).xyz;
+    highp vec3 bump = texture2D(NormalMap, texCoordVarying.st).xyz;
     bump -= vec3(.5);
     bump *= 2.;
     n = normalize(tbn * bump);
     
-    mediump vec3 l = normalize(LightPos - pos).xyz;
+    highp vec3 l = normalize(LightPos - pos).xyz;
     
-    lowp float intensity = dot(n, l);
-    lowp vec3 diffuse = Kd * clamp(intensity, 0.0, 1.0) * Id;
-    lowp vec4 diffuseResult = vec4(clamp(diffuse, 0.0, 1.0), 1.0);
+    highp float intensity = dot(n, l);
+    highp vec3 diffuse = Kd * clamp(intensity, 0.0, 1.0) * Id;
+    highp vec4 diffuseResult = vec4(clamp(diffuse, 0.0, 1.0), 1.0);
     
     // If vertex is lit, calculate specular term in view space using the Blinn-Phong model
-    lowp vec4 specularResult = vec4(0.0);
+    highp vec4 specularResult = vec4(0.0);
     if (intensity > 0.0)
     {
-        mediump vec3 eyeVec = normalize(EyePos.xyz - pos.xyz);
-        mediump vec3 h = normalize(l + eyeVec);
+        highp vec3 eyeVec = normalize(EyePos.xyz - pos.xyz);
+        highp vec3 h = normalize(l + eyeVec);
         
-        mediump float specIntensity = pow(max(dot(h,n), 0.0), Ns);
-        mediump vec3 specular = Ks * clamp(specIntensity, 0.0, 1.0) * Is;
+        highp float specIntensity = pow(max(dot(h,n), 0.0), Ns);
+        highp vec3 specular = Ks * clamp(specIntensity, 0.0, 1.0) * Is;
         specularResult = vec4(clamp(specular, 0.0, 1.0), 1.0)*texture2DProj(SpecularMap, texCoordVarying);
     }
     
-    lowp vec4 color = texture2DProj(DiffuseMap, texCoordVarying);
+    highp vec4 color = texture2DProj(DiffuseMap, texCoordVarying);
     gl_FragColor = /*vec4(0.0, 0.0, 1.0, 1.0);*/ /* (ambientResult + diffuseResult) */ color /*+ specularResult*/;
 }
